- extracted some parts - removed some leftover code from the previous improvements - general code cleanup - improved stability - fixed a bug where the settings were not loaded initially
